requirements.txt
asgiref~=3.8.1
Django~=5.2.2
django-browser-reload~=1.13.0
python-decouple~=3.8
sqlparse~=0.5.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
